requirements.txt
swarms>=5.0.0
openai>=1.0.0
pillow>=10.0.0
pyttsx3>=2.90
gTTS>=2.3.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
twilio>=8.0.0
